Move LM-lookahead from AdvancedTreeSearch to Search
