+++ src/Search/AdvancedTreeSearch/Helpers.hh
@@ -73,21 +73,6 @@ private:
     Core::Statistics<f32>& timeStats_;
 };
 
-inline f32 scaledLogAdd(f32 a, f32 b, f32 scale, f32 invertedScale) {
-    if (b == Core::Type<f32>::max)
-        return a;
-    if (a == Core::Type<f32>::max)
-        return b;
-    a *= invertedScale;
-    b *= invertedScale;
-    return scale * (std::min(a, b) - ::log1p(::exp(std::min(a, b) - std::max(a, b))));
-}
-
-inline bool approximatelyEqual(double a, double b, const double threshold = 0.001) {
-    double diff = a - b;
-    return diff > -threshold && diff < threshold;
-}
-
 // Helper function
 template<class T>
 inline void overwriteOrPush(u32& size, std::vector<T>& vec, const T& value) {
